refactor cpp files
