+++ src/servers/rapier_physics_server_2d.cpp
@@ -1,6 +1,21 @@
 #include "rapier_physics_server_2d.h"
 
-#include "rapier_body_direct_state_2d.h"
+#include "../bodies/rapier_body_direct_state_2d.h"
+
+#include "../spaces/rapier_direct_space_state_2d.h"
+
+#include "../shapes/rapier_world_boundary_shape_2d.h"
+#include "../shapes/rapier_separation_ray_shape_2d.h"
+#include "../shapes/rapier_segment_shape_2d.h"
+#include "../shapes/rapier_circle_shape_2d.h"
+#include "../shapes/rapier_rectangle_shape_2d.h"
+#include "../shapes/rapier_capsule_shape_2d.h"
+#include "../shapes/rapier_convex_polygon_shape_2d.h"
+#include "../shapes/rapier_concave_polygon_shape_2d.h"
+
+#include "../joints/rapier_damped_spring_joint_2d.h"
+#include "../joints/rapier_groove_joint_2d.h"
+#include "../joints/rapier_pin_joint_2d.h"
 
 #define FLUSH_QUERY_CHECK(m_object) \
 	ERR_FAIL_COND_MSG(m_object->get_space() && flushing_queries, "Can't change this state while flushing queries. Use call_deferred() or set_deferred() to change monitoring state instead.");
@@ -968,39 +983,37 @@ void RapierPhysicsServer2D::_joint_clear(const RID &p_joint) {
 }
 
 void RapierPhysicsServer2D::_joint_set_param(const RID &p_joint, JointParam p_param, double p_value) {
-	// RapierJoint2D *joint = joint_owner.get_or_null(p_joint);
-	// ERR_FAIL_COND(!joint);
+	RapierJoint2D *joint = joint_owner.get_or_null(p_joint);
+	ERR_FAIL_COND(!joint);
 
-	// switch (p_param) {
-	// 	case JOINT_PARAM_BIAS:
-	// 		joint->set_bias(p_value);
-	// 		break;
-	// 	case JOINT_PARAM_MAX_BIAS:
-	// 		joint->set_max_bias(p_value);
-	// 		break;
-	// 	case JOINT_PARAM_MAX_FORCE:
-	// 		joint->set_max_force(p_value);
-	// 		break;
-	// }
+	switch (p_param) {
+	case JOINT_PARAM_BIAS:
+		joint->set_bias(p_value);
+	 		break;
+	 	case JOINT_PARAM_MAX_BIAS:
+	 		joint->set_max_bias(p_value);
+	 		break;
+	 	case JOINT_PARAM_MAX_FORCE:
+	 		joint->set_max_force(p_value);
+	 		break;
+	 }
 }
 
 double RapierPhysicsServer2D::_joint_get_param(const RID &p_joint, JointParam p_param) const {
-	// const RapierJoint2D *joint = joint_owner.get_or_null(p_joint);
-	// ERR_FAIL_COND_V(!joint, -1);
+	 const RapierJoint2D *joint = joint_owner.get_or_null(p_joint);
+	 ERR_FAIL_COND_V(!joint, -1);
 
-	// switch (p_param) {
-	// 	case JOINT_PARAM_BIAS:
-	// 		return joint->get_bias();
-	// 		break;
-	// 	case JOINT_PARAM_MAX_BIAS:
-	// 		return joint->get_max_bias();
-	// 		break;
-	// 	case JOINT_PARAM_MAX_FORCE:
-	// 		return joint->get_max_force();
-	// 		break;
-	// }
-
-	return 0;
+	 switch (p_param) {
+	 	case JOINT_PARAM_BIAS:
+	 		return joint->get_bias();
+	 		break;
+	 	case JOINT_PARAM_MAX_BIAS:
+	 		return joint->get_max_bias();
+	 		break;
+	 	case JOINT_PARAM_MAX_FORCE:
+	 		return joint->get_max_force();
+	 		break;
+	 }
 }
 
 void RapierPhysicsServer2D::_joint_disable_collisions_between_bodies(const RID &p_joint, const bool p_disable) {
@@ -1107,22 +1120,21 @@ double RapierPhysicsServer2D::_pin_joint_get_param(const RID &p_joint, PinJointP
 }
 
 void RapierPhysicsServer2D::_damped_spring_joint_set_param(const RID &p_joint, DampedSpringParam p_param, double p_value) {
-	// RapierJoint2D *j = joint_owner.get_or_null(p_joint);
-	// ERR_FAIL_COND(!j);
-	// ERR_FAIL_COND(j->get_type() != JOINT_TYPE_DAMPED_SPRING);
+	RapierJoint2D *j = joint_owner.get_or_null(p_joint);
+	ERR_FAIL_COND(!j);
+	ERR_FAIL_COND(j->get_type() != JOINT_TYPE_DAMPED_SPRING);
 
-	// RapierDampedSpringJoint2D *dsj = static_cast<RapierDampedSpringJoint2D *>(j);
-	// dsj->set_param(p_param, p_value);
+	RapierDampedSpringJoint2D *dsj = static_cast<RapierDampedSpringJoint2D *>(j);
+	dsj->set_param(p_param, p_value);
 }
 
 double RapierPhysicsServer2D::_damped_spring_joint_get_param(const RID &p_joint, DampedSpringParam p_param) const {
-	// RapierJoint2D *j = joint_owner.get_or_null(p_joint);
-	// ERR_FAIL_COND_V(!j, 0);
-	// ERR_FAIL_COND_V(j->get_type() != JOINT_TYPE_DAMPED_SPRING, 0);
+	RapierJoint2D *j = joint_owner.get_or_null(p_joint);
+	ERR_FAIL_COND_V(!j, 0);
+	ERR_FAIL_COND_V(j->get_type() != JOINT_TYPE_DAMPED_SPRING, 0);
 
-	// RapierDampedSpringJoint2D *dsj = static_cast<RapierDampedSpringJoint2D *>(j);
-	// return dsj->get_param(p_param);
-	return 0;
+	RapierDampedSpringJoint2D *dsj = static_cast<RapierDampedSpringJoint2D *>(j);
+	return dsj->get_param(p_param);
 }
 
 PhysicsServer2D::JointType RapierPhysicsServer2D::_joint_get_type(const RID &p_joint) const {
