+++ src/servers/rapier_physics_server_2d.h
@@ -14,11 +14,11 @@
 
 #include "../bodies/rapier_area_2d.h"
 #include "../bodies/rapier_body_2d.h"
-#include "../joints/rapier_joints_2d.h"
+#include "../joints/rapier_joint_2d.h"
 #include "../shapes/rapier_shape_2d.h"
 #include "../spaces/rapier_space_2d.h"
 
-#include "rapier_include.h"
+#include "../rapier_include.h"
 
 using namespace godot;
