+++ src/shapes/rapier_shape_2d.h
@@ -1,10 +1,11 @@
-#pragma once
+#ifndef RAPIER_SHAPE_2D_H
+#define RAPIER_SHAPE_2D_H
 
 #include <godot_cpp/classes/physics_server2d.hpp>
 #include <godot_cpp/templates/hash_map.hpp>
 #include <godot_cpp/templates/local_vector.hpp>
 
-#include "rapier_include.h"
+#include "../rapier_include.h"
 
 using namespace godot;
 
@@ -64,154 +65,4 @@ public:
 	virtual ~RapierShape2D();
 };
 
-class RapierWorldBoundaryShape2D : public RapierShape2D {
-	Vector2 normal;
-	real_t d = 0.0;
-
-protected:
-	virtual rapier2d::Handle create_rapier_shape() const override;
-
-public:
-	virtual PhysicsServer2D::ShapeType get_type() const override { return PhysicsServer2D::SHAPE_WORLD_BOUNDARY; }
-
-	virtual void apply_rapier_transform(rapier2d::Vector &position, real_t &angle) const override;
-
-	virtual void set_data(const Variant &p_data) override;
-	virtual Variant get_data() const override;
-
-	virtual real_t get_moment_of_inertia(real_t p_mass, const Size2 &p_scale) const override { return 0.0; }
-};
-
-class RapierSeparationRayShape2D : public RapierSegmentShape2D {
-	real_t length = 0.0;
-	bool slide_on_slope = false;
-
-protected:
-	virtual rapier2d::Handle create_rapier_shape() const override;
-public:
-	virtual PhysicsServer2D::ShapeType get_type() const override { return PhysicsServer2D::SHAPE_SEPARATION_RAY; }
-
-	virtual void apply_rapier_transform(rapier2d::Vector &position, real_t &angle) const override;
-
-	virtual void set_data(const Variant &p_data) override;
-	virtual Variant get_data() const override;
-
-	virtual real_t get_moment_of_inertia(real_t p_mass, const Size2 &p_scale) const override;
-};
-
-class RapierSegmentShape2D : public RapierShape2D {
-	Vector2 a;
-	Vector2 b;
-	Vector2 n;
-
-protected:
-	virtual rapier2d::Handle create_rapier_shape() const override;
-
-public:
-	_FORCE_INLINE_ const Vector2 &get_a() const { return a; }
-	_FORCE_INLINE_ const Vector2 &get_b() const { return b; }
-	_FORCE_INLINE_ const Vector2 &get_normal() const { return n; }
-
-	virtual PhysicsServer2D::ShapeType get_type() const override { return PhysicsServer2D::SHAPE_SEGMENT; }
-
-	virtual void set_data(const Variant &p_data) override;
-	virtual Variant get_data() const override;
-
-	virtual real_t get_moment_of_inertia(real_t p_mass, const Size2 &p_scale) const override;
-};
-
-class RapierCircleShape2D : public RapierShape2D {
-	real_t radius;
-
-protected:
-	virtual rapier2d::Handle create_rapier_shape() const override;
-
-public:
-	_FORCE_INLINE_ const real_t &get_radius() const { return radius; }
-
-	virtual PhysicsServer2D::ShapeType get_type() const override { return PhysicsServer2D::SHAPE_CIRCLE; }
-
-	virtual void set_data(const Variant &p_data) override;
-	virtual Variant get_data() const override;
-
-	virtual real_t get_moment_of_inertia(real_t p_mass, const Size2 &p_scale) const override;
-};
-
-class RapierRectangleShape2D : public RapierShape2D {
-	Vector2 half_extents;
-
-protected:
-	virtual rapier2d::Handle create_rapier_shape() const override;
-
-public:
-	_FORCE_INLINE_ const Vector2 &get_half_extents() const { return half_extents; }
-
-	virtual PhysicsServer2D::ShapeType get_type() const override { return PhysicsServer2D::SHAPE_RECTANGLE; }
-
-	virtual void set_data(const Variant &p_data) override;
-	virtual Variant get_data() const override;
-
-	virtual real_t get_moment_of_inertia(real_t p_mass, const Size2 &p_scale) const override;
-};
-
-class RapierConvexPolygonShape2D : public RapierShape2D {
-	struct Point {
-		Vector2 pos;
-		Vector2 normal; //normal to next segment
-	};
-
-	Point *points = nullptr;
-	int point_count = 0;
-
-public:
-	virtual PhysicsServer2D::ShapeType get_type() const override { return PhysicsServer2D::SHAPE_CONVEX_POLYGON; }
-
-	virtual void set_data(const Variant &p_data) override;
-	virtual Variant get_data() const override;
-
-	virtual real_t get_moment_of_inertia(real_t p_mass, const Size2 &p_scale) const override;
-
-	~RapierConvexPolygonShape2D();
-
-protected:
-	virtual rapier2d::Handle create_rapier_shape() const override;
-};
-
-class RapierCapsuleShape2D : public RapierShape2D {
-	real_t radius = 0.0;
-	real_t height = 0.0;
-
-public:
-	_FORCE_INLINE_ const real_t &get_radius() const { return radius; }
-	_FORCE_INLINE_ const real_t &get_height() const { return height; }
-
-	virtual PhysicsServer2D::ShapeType get_type() const override { return PhysicsServer2D::SHAPE_CAPSULE; }
-
-	virtual void set_data(const Variant &p_data) override;
-	virtual Variant get_data() const override;
-
-	virtual real_t get_moment_of_inertia(real_t p_mass, const Size2 &p_scale) const override;
-
-protected:
-	virtual rapier2d::Handle create_rapier_shape() const override;
-};
-
-class RapierConcavePolygonShape2D : public RapierShape2D {
-	struct Segment {
-		int points[2] = {};
-	};
-
-	LocalVector<Segment> segments;
-	LocalVector<Point2> points;
-
-protected:
-	virtual rapier2d::Handle create_rapier_shape() const override;
-
-public:
-	virtual PhysicsServer2D::ShapeType get_type() const override { return PhysicsServer2D::SHAPE_CONCAVE_POLYGON; }
-
-	virtual void set_data(const Variant &p_data) override;
-	virtual Variant get_data() const override;
-
-	virtual real_t get_moment_of_inertia(real_t p_mass, const Size2 &p_scale) const override { return 0.0; }
-};
+#endif // RAPIER_SHAPE_2D_H
