+++ src/spaces/rapier_direct_space_state_2d.h
@@ -1,8 +1,9 @@
-#ifndef RAPIER_SPACE_2D_H
-#define RAPIER_SPACE_2D_H
+#ifndef RAPIER_DIRECT_SPACE_STATE_2D_H
+#define RAPIER_DIRECT_SPACE_STATE_2D_H
 
-#include "rapier_area_2d.h"
-#include "rapier_body_2d.h"
+#include "../bodies/rapier_area_2d.h"
+#include "../bodies/rapier_body_2d.h"
+#include "rapier_space_2d.h"
 
 #include <gdextension_interface.h>
 #include <godot_cpp/classes/physics_direct_space_state2d.hpp>
@@ -34,3 +35,5 @@ public:
 
 	RapierDirectSpaceState2D() {}
 };
+
+#endif // RAPIER_DIRECT_SPACE_STATE_2D_H
